Replace BucketInput with NumberInput WIP
